适应 ib 修改 x、y、w、h、speed 等为 double，修改 acc 参数为 double 类型，修复 Fish 移动 bug
1.适应 ib 修改 x、y、w、h、speed 等为 double
==>各 IBCharacter 子类构造方法参数改为 double 类型
==>去掉一些不再需要的取整强制转换

2.修改 acc 参数为 double 类型

3.修复 Fish 移动 bug
==>初始化 Fish 计算 speed 为 0，出现加速度为 NaN 的情况
==>之前 speed 为 int 类型， NaN 强转为 int 结果为 0，导致没有发现问题
==>增加不为 0 的判断

+++ src/main/java/cn/milai/ibdemo/character/EscapeCraft.java
@@ -10,7 +10,7 @@ import cn.milai.ib.container.ui.UIContainer;
  */
 public class EscapeCraft extends AbstractIBCharacter {
 
-	public EscapeCraft(int x, int y, UIContainer container) {
+	public EscapeCraft(double x, double y, UIContainer container) {
 		super(x, y, container);
 	}
