+++ src/main/java/cn/milai/ibdemo/character/UltraFly.java
@@ -3,9 +3,13 @@ package cn.milai.ibdemo.character;
 import cn.milai.ib.character.AbstractIBCharacter;
 import cn.milai.ib.container.ui.UIContainer;
 
+/**
+ * 飞行乌托拉
+ * @author milai
+ */
 public class UltraFly extends AbstractIBCharacter {
 
-	public UltraFly(int x, int y, UIContainer container) {
+	public UltraFly(double x, double y, UIContainer container) {
 		super(x, y, container);
 	}
